backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
asyncio==3.4.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cod-api==2.0.1
DateTime==5.3
Django==4.2.7
djangorestframework==3.14.0
enum34==1.1.10
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg2==2.9.9
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
uuid==1.30
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.1
